method requestData moved to class EnterData
