+++ src/service/HorizontalService.java
@@ -1,8 +1,6 @@
 package service;
 
-import model.Blind;
-
 public interface HorizontalService {
 
-    long calculateCost(Blind blind, PriceCatalog priceCatalogImpl);
+    long calculateCost();
 }
