+++ src/service/PriceCatalog.java
@@ -5,5 +5,6 @@ import java.util.Map;
 public interface PriceCatalog {
 
     Map<Integer, Double> initDate();
+
     double getColorPrice(int color);
 }
