+++ src/view/EnterData.java
@@ -1,0 +1,8 @@
+package view;
+
+import model.Blind;
+
+public interface EnterData {
+
+    Blind requestData();
+}
