+++ src/view/UserPage.java
@@ -1,9 +1,7 @@
 package view;
 
-import model.Blind;
-
 public interface UserPage {
 
     void showBlindCost();
-    Blind requestData();
+
 }
