+++ src/view/UserPageImpl.java
@@ -1,6 +1,5 @@
 package view;
 
-import model.Blind;
 import service.HorizontalService;
 import service.HorizontalServiceImpl;
 import service.PriceCatalog;
@@ -8,10 +7,12 @@ import service.PriceCatalogImpl;
 
 import java.util.Scanner;
 
-public class UserPageImpl implements UserPage{
+public class UserPageImpl implements UserPage {
 
-    HorizontalService horizontalServiceImpl = new HorizontalServiceImpl();
+    EnterData enterData = new EnterDataImpl();
     PriceCatalog priceCatalogImpl = new PriceCatalogImpl();
+    HorizontalService horizontalServiceImpl;
+
     Scanner scanner = new Scanner(System.in);
     Scanner scannerStr = new Scanner(System.in);
     long costBlinds;
@@ -23,7 +24,8 @@ public class UserPageImpl implements UserPage{
             System.out.println("Would you like to calculate the cost of blinds?\nIf yes - enter Y, if no - enter N");
             String answer = scannerStr.nextLine();
             if (answer.equalsIgnoreCase("Y")) {
-                costBlinds = horizontalServiceImpl.calculateCost(requestData(), priceCatalogImpl);
+                horizontalServiceImpl = new HorizontalServiceImpl(enterData.requestData(), priceCatalogImpl);
+                costBlinds = horizontalServiceImpl.calculateCost();
                 System.out.println("model.Blind costs " + costBlinds + " rubles.\n");
             } else if (answer.equalsIgnoreCase("N")) {
                 System.out.println("Calculation finished.");
@@ -34,15 +36,4 @@ public class UserPageImpl implements UserPage{
         }
     }
 
-    @Override
-    public Blind requestData() {
-        Blind blind = new Blind();
-        System.out.println("Enter the width of the blind in mm: ");
-        blind.setWidth(scanner.nextInt());
-        System.out.println("Enter the height of the blind in mm:");
-        blind.setHeight(scanner.nextInt());
-        System.out.println("Enter the color number of the blind (201, 202): ");
-        blind.setColor(scanner.nextInt());
-        return blind;
-    }
 }
